Hey, Just added starting files.
Working on merge to better class organization.  I am working on a
"beads" class that will manage the different beads.  There is also a
bead class and a bin class.  The bin objects will all coraspond to a
bin, and the bead objects will all coraspond to beads.  Bin objects will
have beads.  The beads class might be used for reading the bead info
from a .bin file in the future.